t2893: extract FAIL_REASON + formatGateThrowMessage to sibling module
Drops quality-hooks-signature.mjs total complexity below the qlty
'High total complexity' regression threshold. The companion module
holds the structured failure-reason enum and the throw-message
formatter so the parent file stays under the per-file complexity
ratchet.

FAIL_REASON is re-exported from the parent so existing test imports
keep working without churn.

Also resolves a stale TODO.md merge conflict (duplicate long-form
entries for t2871/t2872/t2877/t2888 — canonical short forms exist
elsewhere in the file).

For #21030
