+++ .agents/plugins/opencode-aidevops/quality-hooks-signature.mjs
@@ -47,25 +47,15 @@ import { existsSync, readFileSync, appendFileSync } from "fs";
 import { execSync } from "child_process";
 import { join } from "path";
 
+import { FAIL_REASON, formatGateThrowMessage } from "./quality-hooks-signature-failures.mjs";
+
 export const SIG_MARKER = "<!-- aidevops:sig -->";
 
-/**
- * Structured failure reasons for signature repair (t2893).
- *
- * Each reason names a specific failure mode the repair pipeline can hit.
- * The throw site formats the matching message into the mentoring error so
- * the next attempt knows precisely what to fix instead of guessing across
- * three unrelated causes.
- */
-export const FAIL_REASON = {
-  FILE_NOT_FOUND: "body-file not found (may be created later in this same bash call)",
-  FILE_UNREADABLE: "body-file exists but cannot be read",
-  HELPER_MISSING: "gh-signature-helper.sh not found",
-  HELPER_FAILED: "gh-signature-helper.sh invocation failed",
-  UNPARSEABLE_BODY: "body uses heredoc, process substitution, or command substitution",
-  BODY_ARG_QUOTING: "signature contains delimiter quote, cannot safely rewrite --body",
-  BODY_ARG_NO_MATCH: "could not parse --body argument shape",
-};
+// Re-export FAIL_REASON so existing test imports
+// (`import { FAIL_REASON } from "../quality-hooks-signature.mjs"`) keep
+// working. Definition lives in quality-hooks-signature-failures.mjs to
+// keep this file under the qlty per-file complexity threshold (t2893).
+export { FAIL_REASON };
 
 /**
  * Strip heredoc body lines from a multi-line command string.
@@ -389,52 +379,6 @@ export function tryRepairSignature(cmd, scriptsDir, log) {
 }
 
 /**
- * Build the throw error message for a structured repair failure (t2893).
- *
- * Adds targeted same-bash-call mentorship for the FILE_NOT_FOUND case —
- * the most common shape that tripped the pre-t2893 generic message.
- * @param {{ reason: string, detail?: string }} failure
- * @param {string} cmdSnippet
- * @returns {string}
- */
-function _formatGateThrowMessage(failure, cmdSnippet) {
-  const causeBlock = failure.detail
-    ? `Specific cause: ${failure.reason} (${failure.detail})`
-    : `Specific cause: ${failure.reason}`;
-
-  const sameCallHint =
-    failure.reason === FAIL_REASON.FILE_NOT_FOUND
-      ? "\n\nLikely cause: the body-file is created in this same bash call. " +
-        "This hook runs PRE-execution, so it cannot see files that bash " +
-        "hasn't created yet. Two correct fixes:\n" +
-        "  a. Split into two bash tool calls: one to create the file, one " +
-        "to gh issue/pr comment.\n" +
-        "  b. Source shared-gh-wrappers.sh and call gh_issue_comment / " +
-        "gh_create_pr / gh_create_issue / gh_pr_comment by name. The shell " +
-        "wrapper runs AFTER the file-creation steps in the same bash call, " +
-        "so it sees the completed file. The PATH shim at .agents/scripts/gh " +
-        "also runs at exec-time as a backstop."
-      : "";
-
-  return (
-    `aidevops: gh write command blocked at signature gate (t2685, t2893).\n\n` +
-    `${causeBlock}${sameCallHint}\n\n` +
-    `Standard fixes:\n` +
-    `  1. Append to --body directly:\n` +
-    `       gh issue comment N --body "...$(gh-signature-helper.sh footer)"\n` +
-    `  2. Append to --body-file (two-step pattern):\n` +
-    `       gh-signature-helper.sh footer >> "$BODY_FILE"\n` +
-    `       gh issue comment N --body-file "$BODY_FILE"\n` +
-    `  3. Source the wrapper and call by name:\n` +
-    `       source ~/.aidevops/agents/scripts/shared-gh-wrappers.sh\n` +
-    `       gh_issue_comment N --body-file "$BODY_FILE"\n\n` +
-    `Cmd snippet: ${cmdSnippet}\n\n` +
-    `Emergency bypass (last resort, breaks audit trail): set\n` +
-    `AIDEVOPS_GH_SHIM_DISABLE=1 — but this plugin hook still blocks.`
-  );
-}
-
-/**
  * Check signature footer gate on gh write commands (GH#12805, t1755, t2685, t2893).
  * @param {string} cmd - Bash command string
  * @param {Function} log - Quality logger function
@@ -470,6 +414,6 @@ export function checkSignatureFooterGate(cmd, log, scriptsDir, output) {
         (result.detail ? ` (${result.detail})` : "") +
         ` | cmd: ${snippet}`,
     );
-    throw new Error(_formatGateThrowMessage(result, snippet));
+    throw new Error(formatGateThrowMessage(result, snippet));
   }
 }
